// web/src/api/generated/models/scraper-task-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ScraperTaskResponse {
    'id'?: number;
    'configId'?: number;
    'configName'?: string;
    'status'?: string;
    'statusDescription'?: string;
    'totalArticles'?: number;
    'successCount'?: number;
    'failCount'?: number;
    'createdArticleIds'?: Array<number>;
    'errorMessage'?: string;
    'startTime'?: string;
    'endTime'?: string;
    'durationMs'?: number;
    'createTime'?: string;
}