// web/src/api/generated/models/scraper-task-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ScraperTaskResponse } from './scraper-task-response';

/**
 * 抓取任务分页响应
 */
export interface ScraperTaskPageResponse {
    /**
     * 任务列表
     */
    'records'?: Array<ScraperTaskResponse>;
    /**
     * 总记录数
     */
    'total'?: number;
    /**
     * 当前页码
     */
    'pageNum'?: number;
    /**
     * 每页大小
     */
    'pageSize'?: number;
    /**
     * 总页数
     */
    'totalPages'?: number;
}