// web/src/api/generated/models/scraper-config-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ScraperConfigResponse {
    'id'?: number;
    'name'?: string;
    'sourceCode'?: string;
    'baseUrl'?: string;
    'description'?: string;
    'titleSelector'?: string;
    'authorSelector'?: string;
    'sourceSelector'?: string;
    'contentSelector'?: string;
    'dateSelector'?: string;
    'imageSelector'?: string;
    'linkSelector'?: string;
    'maxDepth'?: number;
    'maxPages'?: number;
    'delayMs'?: number;
    'useDynamic'?: boolean;
    'waitForJsMs'?: number;
    'cronExpression'?: string;
    'enabled'?: boolean;
    'defaultMaxArticles'?: number;
    'defaultCategory'?: string;
    'defaultDifficulty'?: number;
    'creatorId'?: number;
    'createTime'?: string;
    'updateTime'?: string;
}